Re-wire prometheus integration
