+++ Cargo.toml
@@ -52,6 +52,7 @@ lazy_static = "1.4.0"
 url = "2.1.0"
 percent-encoding = "2.1.0"
 sha2 = "0.8.0"
+prometheus = { version = "0.9", optional = true }
 
 [dependencies.serde_json]
 version = "1.0"
@@ -84,6 +85,7 @@ criterion = "0.3"
 [features]
 developer-mode = []
 default = ["developer-mode"]
+monitoring = ["prometheus"]
 
 [target.'cfg(all(target_arch = "x86_64", not(target_env = "msvc")))'.dependencies]
 sha2-asm = "0.5.3"
