+++ testnet/stacks-node/src/run_loop/mod.rs
@@ -1,9 +1,6 @@
 pub mod helium;
 pub mod neon;
 
-#[cfg(feature = "monitoring")]
-pub mod monitoring;
-
 use crate::{BurnchainController, BurnchainTip, ChainTip, Tenure};
 
 use stacks::chainstate::stacks::{TransactionAuth, TransactionSpendingCondition, TransactionPayload};
